+++ CHANGELOG.md
@@ -7,7 +7,12 @@
 * Can flip sprites and text vertically and horizontally.
 * Save data manager added to handle saving and loading saves.
     * Currenty only works on Windows.
-* Projects are now autosaved at a regular interval.
+* Projects are now auto-saved at a regular interval.
+* Crashes are handled more gracefully.
+    * If there is an error, it will be logged.
+	* Settings will be saved, if possible.
+	* The current project will be auto-saved if possible.
+	    * This will not overwrite any existing auto-saves.
 
 ## Refactors